code cleanup and optimization
